requirements.txt
pip==21.1.3
fastapi==0.66.0
pydantic==1.8.2
Jinja2==3.0.1
pathlib==1.0.1
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
starlette==0.14.2
deta==1.0.0
bcrypt==3.2.0
pydantic[email]
hurry.filesize==0.9
aiofiles
shortuuid
uvicorn
